refactor(auth): split storage and account helpers
