+++ src/notebooklm/auth.py
@@ -29,8 +29,6 @@ Security Notes:
 
 import asyncio
 import contextlib
-import errno
-import json
 import logging
 import os
 import re
@@ -39,22 +37,20 @@ import subprocess
 import sys
 import threading
 import time
-import warnings
 import weakref
 from collections.abc import Iterator
 from contextvars import ContextVar
 from dataclasses import dataclass
 from pathlib import Path
 from types import ModuleType
-from typing import Any, NamedTuple, TypeAlias
-from urllib.parse import urlencode
+from typing import Any, TypeAlias
 
 import httpx
-from filelock import FileLock
 
-from ._atomic_io import atomic_update_json, atomic_write_json
+from ._auth import account as _auth_account
 from ._auth import cookie_policy as _cookie_policy
 from ._auth import cookies as _auth_cookies
+from ._auth import storage as _auth_storage
 from ._env import get_base_url
 from ._url_utils import contains_google_auth_redirect, is_google_auth_redirect
 from .exceptions import AuthExtractionError
@@ -86,49 +82,21 @@ load_httpx_cookies = _auth_cookies.load_httpx_cookies
 normalize_cookie_map = _auth_cookies.normalize_cookie_map
 
 
-class CookieSnapshotKey(NamedTuple):
-    """Path-aware cookie identity used by the snapshot/delta save machinery.
-
-    RFC 6265 treats ``path`` as part of cookie identity: two cookies with the
-    same ``(name, domain)`` but different paths are distinct entries. The
-    snapshot/delta path widens the legacy ``(name, domain)`` key (still used
-    elsewhere for back-compat — see ``CookieKey``) to ``(name, domain, path)``
-    so that path-scoped cookies (e.g. ``OSID`` on a per-product path) survive
-    a load → save round trip and so that a sibling-process write to a
-    different-path variant of the same name is not silently overwritten.
-    """
-
-    name: str
-    domain: str
-    path: str
-
-
-class CookieSnapshotValue(NamedTuple):
-    """Snapshot value tuple: ``(value, expires, secure, http_only)``.
-
-    Widened from a bare ``str`` so that a ``Set-Cookie`` which keeps the same
-    value but renews ``expires`` (or flips ``secure`` / ``httpOnly``) still
-    registers as a delta. The legacy save path compared ``expires`` directly
-    and would write the new expiry through; the snapshot path previously
-    keyed on value alone and silently dropped attribute-only refreshes.
-    """
-
-    value: str
-    expires: int | None
-    secure: bool
-    http_only: bool
-
-
-CookieSnapshot: TypeAlias = dict[CookieSnapshotKey, CookieSnapshotValue]
-
-
-@dataclass(frozen=True)
-class CookieSaveResult:
-    """Detailed result for callers that need to maintain a save baseline."""
-
-    ok: bool
-    cas_rejected_keys: frozenset[CookieSnapshotKey] = frozenset()
-
+CookieSnapshotKey = _auth_storage.CookieSnapshotKey
+CookieSnapshotValue = _auth_storage.CookieSnapshotValue
+CookieSnapshot = _auth_storage.CookieSnapshot
+CookieSaveResult = _auth_storage.CookieSaveResult
+snapshot_cookie_jar = _auth_storage.snapshot_cookie_jar
+advance_cookie_snapshot_after_save = _auth_storage.advance_cookie_snapshot_after_save
+_cookie_save_return = _auth_storage._cookie_save_return
+save_cookies_to_storage = _auth_storage.save_cookies_to_storage
+_merge_cookies_legacy = _auth_storage._merge_cookies_legacy
+_merge_cookies_with_snapshot = _auth_storage._merge_cookies_with_snapshot
+_cookie_snapshot_key_variants = _auth_storage._cookie_snapshot_key_variants
+_stored_cookie_snapshot_key = _auth_storage._stored_cookie_snapshot_key
+_file_lock = _auth_storage._file_lock
+_file_lock_exclusive = _auth_storage._file_lock_exclusive
+_FLOCK_UNAVAILABLE_WARNED = _auth_storage._FLOCK_UNAVAILABLE_WARNED
 
 REQUIRED_COOKIE_DOMAINS = _cookie_policy.REQUIRED_COOKIE_DOMAINS
 OPTIONAL_COOKIE_DOMAINS_BY_LABEL = _cookie_policy.OPTIONAL_COOKIE_DOMAINS_BY_LABEL
@@ -145,11 +113,57 @@ _is_allowed_auth_domain = _cookie_policy._is_allowed_auth_domain
 _is_allowed_cookie_domain = _cookie_policy._is_allowed_cookie_domain
 
 
+_AUTH_STORAGE_FACADE_NAMES = {
+    "CookieSnapshotKey",
+    "CookieSnapshotValue",
+    "CookieSnapshot",
+    "CookieSaveResult",
+    "snapshot_cookie_jar",
+    "advance_cookie_snapshot_after_save",
+    "_cookie_save_return",
+    "save_cookies_to_storage",
+    "_merge_cookies_legacy",
+    "_merge_cookies_with_snapshot",
+    "_cookie_snapshot_key_variants",
+    "_stored_cookie_snapshot_key",
+    "_file_lock",
+    "_file_lock_exclusive",
+    "_FLOCK_UNAVAILABLE_WARNED",
+}
+
+_AUTH_ACCOUNT_FACADE_NAMES = {
+    "Account",
+    "MAX_AUTHUSER_PROBE",
+    "_ACCOUNT_CONTEXT_KEY",
+    "_account_context_path",
+    "extract_email_from_html",
+    "_probe_authuser",
+    "read_account_metadata",
+    "get_authuser_for_storage",
+    "get_account_email_for_storage",
+    "format_authuser_value",
+    "authuser_query",
+    "write_account_metadata",
+    "clear_account_metadata",
+}
+
+
 class _AuthFacadeModule(ModuleType):
     """Keep compatibility assignments to auth globals meaningful after moves."""
 
+    def __getattribute__(self, name: str) -> Any:
+        if name in _AUTH_STORAGE_FACADE_NAMES:
+            return getattr(_auth_storage, name)
+        if name in _AUTH_ACCOUNT_FACADE_NAMES:
+            return getattr(_auth_account, name)
+        return super().__getattribute__(name)
+
     def __setattr__(self, name: str, value: Any) -> None:
         super().__setattr__(name, value)
+        if name in _AUTH_STORAGE_FACADE_NAMES:
+            setattr(_auth_storage, name, value)
+        if name in _AUTH_ACCOUNT_FACADE_NAMES:
+            setattr(_auth_account, name, value)
         if name == "_SECONDARY_BINDING_WARNED":
             _cookie_policy._SECONDARY_BINDING_WARNED = bool(value)
         elif name in {
@@ -510,327 +524,30 @@ def extract_session_id_from_html(html: str, final_url: str = "") -> str:
     )
 
 
-@dataclass(frozen=True)
-class Account:
-    """A Google account discovered via authuser=N probing.
-
-    Attributes:
-        authuser: The integer index used in ``?authuser=N`` URL parameters.
-            Index 0 is the default account; subsequent indices follow the
-            order Google reports for the browser session.
-        email: The account's email address as it appears in the NotebookLM
-            page's ``WIZ_global_data`` block.
-        is_default: True only for the account at ``authuser=0``.
-        browser_profile: For Chromium-family browsers with multiple
-            user-data profiles, the on-disk directory name (``"Default"``,
-            ``"Profile 1"``) the cookies came from. ``None`` for non-chromium
-            browsers and for the legacy single-jar path where source isn't
-            tracked.
-    """
-
-    authuser: int
-    email: str
-    is_default: bool
-    browser_profile: str | None = None
-
-
-# Hard cap on how many ``authuser`` indices to probe before giving up.
-# Google supports up to ~10 simultaneously signed-in accounts in a browser
-# session; ten covers every realistic case and bounds the worst-case probe.
-MAX_AUTHUSER_PROBE = 10
-
-# Local-parts of well-known non-user emails that NotebookLM may embed in page
-# chrome (footer links, support contacts) and must not be misread as the
-# active account. Combined with ``_NON_USER_EMAIL_DOMAINS`` so we only drop
-# the address when *both* match — otherwise legitimate Workspace users like
-# ``support@customer.com`` would be filtered out.
-_NON_USER_EMAIL_LOCALS = frozenset(
-    {
-        "abuse",
-        "feedback",
-        "info",
-        "mail-noreply",
-        "googlemail-noreply",
-        "no-reply",
-        "noreply",
-        "press",
-        "privacy",
-        "support",
-    }
-)
-_NON_USER_EMAIL_DOMAINS = frozenset({"google.com", "accounts.google.com", "gmail.com"})
-
-# Match a quoted email address, e.g. ``"alice@example.com"``. Mirrors how
-# emails appear in the page's WIZ_global_data JSON.
-_EMAIL_RE = re.compile(r'"([A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,})"')
-
-
-def extract_email_from_html(html: str) -> str | None:
-    """Extract the active user's email from a NotebookLM page response.
-
-    Returns the first plausible Google account email found in the HTML,
-    skipping addresses that look like Google's own contact endpoints
-    (e.g. ``support@google.com``, ``noreply@accounts.google.com``).
-
-    Args:
-        html: Page HTML from ``notebooklm.google.com/?authuser=N``.
-
-    Returns:
-        The account's email, or ``None`` if no plausible address was found
-        (typically because the response was a login redirect or the page
-        structure changed).
-    """
-    for match in _EMAIL_RE.finditer(html):
-        email = match.group(1)
-        local, _, domain = email.partition("@")
-        if local.lower() in _NON_USER_EMAIL_LOCALS and domain.lower() in _NON_USER_EMAIL_DOMAINS:
-            continue
-        return email
-    return None
-
-
-# Chromium-style User-Agent for ``enumerate_accounts``. Without a real-browser
-# UA, Google serves a stripped-down page that omits the WIZ_global_data block
-# (and therefore the active user's email), and ``extract_email_from_html``
-# returns None — looking like "no signed-in account". Empirically validated
-# against ``notebooklm.google.com/?authuser=N``.
-_BROWSER_UA = (
-    "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) "
-    "AppleWebKit/537.36 (KHTML, like Gecko) Chrome/138.0.0.0 Safari/537.36"
-)
-
-
-async def _probe_authuser(client: httpx.AsyncClient, n: int) -> str | None:
-    """Probe one ``authuser`` index and return the active email or ``None``.
-
-    Returns ``None`` for auth-redirect or unparseable responses; lets the
-    caller decide whether that means "past the last account" or a real error.
-    HTTP transport errors propagate.
-
-    Only checks the *final* URL for an auth redirect. The page body is not
-    scanned because a healthy NotebookLM page legitimately contains many
-    ``accounts.google.com`` links (account chooser, manage-account menu)
-    that would fool ``contains_google_auth_redirect``.
-    """
-    response = await client.get(
-        f"{get_base_url()}/?{authuser_query(n)}",
-        headers={"User-Agent": _BROWSER_UA, "Accept": "text/html,*/*"},
-    )
-    if response.status_code != 200:
-        return None
-    if is_google_auth_redirect(str(response.url)):
-        return None
-    return extract_email_from_html(response.text)
+Account = _auth_account.Account
+MAX_AUTHUSER_PROBE = _auth_account.MAX_AUTHUSER_PROBE
+_ACCOUNT_CONTEXT_KEY = _auth_account._ACCOUNT_CONTEXT_KEY
+_account_context_path = _auth_account._account_context_path
+extract_email_from_html = _auth_account.extract_email_from_html
+_probe_authuser = _auth_account._probe_authuser
+read_account_metadata = _auth_account.read_account_metadata
+get_authuser_for_storage = _auth_account.get_authuser_for_storage
+get_account_email_for_storage = _auth_account.get_account_email_for_storage
+format_authuser_value = _auth_account.format_authuser_value
+authuser_query = _auth_account.authuser_query
+write_account_metadata = _auth_account.write_account_metadata
+clear_account_metadata = _auth_account.clear_account_metadata
 
 
 async def enumerate_accounts(
     cookie_jar: httpx.Cookies, *, max_authuser: int = MAX_AUTHUSER_PROBE
 ) -> list[Account]:
-    """Enumerate Google accounts visible to the given cookie jar.
-
-    Probes ``https://notebooklm.google.com/?authuser=N`` for ``N`` in
-    ``0..max_authuser`` and parses the active user's email from each response.
-
-    Stop condition: when the email at index ``N>0`` matches the email at
-    index 0, Google has silently fallen back to the default account, meaning
-    ``N`` is past the real count. Without this check the caller would record
-    duplicate phantom accounts; Google does not redirect to login in this
-    case.
-
-    Args:
-        cookie_jar: ``httpx.Cookies`` jar with auth cookies. Not mutated.
-        max_authuser: Hard cap on indices probed (default
-            :data:`MAX_AUTHUSER_PROBE`).
-
-    Returns:
-        Accounts ordered by ``authuser`` index. ``is_default`` is true for
-        index 0 only.
-
-    Raises:
-        ValueError: If ``authuser=0`` itself does not return a signed-in
-            account (cookies expired or invalid).
-        httpx.HTTPError: If the HTTP transport fails.
-    """
-    async with httpx.AsyncClient(
-        cookies=cookie_jar,
-        follow_redirects=True,
-        timeout=httpx.Timeout(10.0, read=60.0),
-    ) as client:
-        # The browser's on-disk cookie DB rotates ``__Secure-1PSIDTS`` every
-        # few minutes, but only when Chrome itself is actively running. A
-        # ``--browser-cookies`` extraction against an idle Chrome lands here
-        # with a stale SIDTS — the SID is fine, but ``notebooklm.google.com``
-        # responds with a redirect to ``accounts.google.com`` and we'd
-        # incorrectly conclude the user is signed out. Poke once to fetch
-        # fresh SIDTS via Set-Cookie before the probes start.
-        await _poke_session(client, None)
-        default_email = await _probe_authuser(client, 0)
-        if default_email is None:
-            raise ValueError(
-                "Authentication expired or invalid; "
-                "authuser=0 did not return a signed-in account. "
-                "Run 'notebooklm login' to re-authenticate."
-            )
-        accounts = [Account(authuser=0, email=default_email, is_default=True)]
-        for n in range(1, max_authuser + 1):
-            email = await _probe_authuser(client, n)
-            if email is None or email == default_email:
-                break
-            accounts.append(Account(authuser=n, email=email, is_default=False))
-        return accounts
-
-
-_ACCOUNT_CONTEXT_KEY = "account"
-
-
-def _account_context_path(storage_path: Path) -> Path:
-    """Return the context.json path that annotates ``storage_path``."""
-    return storage_path.with_name("context.json")
-
-
-def read_account_metadata(storage_path: Path | None) -> dict[str, Any]:
-    """Read profile account metadata from ``context.json``.
-
-    The ``account`` object records the Google ``authuser`` index used when
-    the profile was authenticated. Profiles from before this feature shipped
-    (and profiles for users with a single Google account) have no account
-    metadata and use ``authuser=0``.
-
-    Args:
-        storage_path: Path to ``storage_state.json``. The sibling
-            ``context.json`` stores account metadata. ``None`` means the
-            profile is loaded from ``NOTEBOOKLM_AUTH_JSON``.
-
-    Returns:
-        Parsed metadata dict, or ``{}`` if the file is missing, unreadable,
-        or malformed. Callers should treat a missing ``authuser`` key as 0.
-    """
-    if storage_path is None:
-        return {}
-    context_path = _account_context_path(storage_path)
-    if not context_path.exists():
-        return {}
-    try:
-        data = json.loads(context_path.read_text(encoding="utf-8"))
-    except (OSError, json.JSONDecodeError) as e:
-        logger.debug("account metadata read failed at %s: %s", context_path, e)
-        return {}
-    if not isinstance(data, dict):
-        return {}
-    account = data.get(_ACCOUNT_CONTEXT_KEY)
-    return account if isinstance(account, dict) else {}
-
-
-def get_authuser_for_storage(storage_path: Path | None) -> int:
-    """Return the ``authuser`` index recorded for a profile, defaulting to 0.
-
-    Profiles without account metadata (legacy single-account installs and
-    fresh logins that never set an authuser) are treated as ``authuser=0``,
-    preserving existing behavior.
-
-    Returns:
-        Non-negative ``authuser`` index. Malformed values fall back to 0.
-    """
-    raw = read_account_metadata(storage_path).get("authuser")
-    if isinstance(raw, int) and raw >= 0:
-        return raw
-    return 0
-
-
-def get_account_email_for_storage(storage_path: Path | None) -> str | None:
-    """Return the persisted account email for stable routing, if available."""
-    raw = read_account_metadata(storage_path).get("email")
-    if isinstance(raw, str):
-        email = raw.strip()
-        if email:
-            return email
-    return None
-
-
-def format_authuser_value(authuser: int = 0, account_email: str | None = None) -> str:
-    """Return the explicit NotebookLM auth routing value.
-
-    Google accepts either an integer account index or the account email in the
-    ``authuser`` field. Email is stable across browser account reordering, so it
-    wins when available; otherwise callers retain the existing integer behavior.
-    """
-    if account_email:
-        stripped = account_email.strip()
-        if stripped:
-            return stripped
-    return str(authuser)
-
-
-def authuser_query(authuser: int = 0, account_email: str | None = None) -> str:
-    """Return a URL-encoded ``authuser=...`` query string."""
-    return urlencode({"authuser": format_authuser_value(authuser, account_email)})
-
-
-def write_account_metadata(storage_path: Path, *, authuser: int, email: str | None = None) -> None:
-    """Persist profile account metadata inside sibling ``context.json``.
-
-    Uses :func:`atomic_update_json` so concurrent CLI invocations (e.g., a
-    ``login`` while ``use`` is in flight) cannot lose updates by writing
-    stale snapshots of ``context.json`` over each other.
-
-    Args:
-        storage_path: Path to ``storage_state.json``. The sibling
-            ``context.json`` is created or updated.
-        authuser: ``authuser`` index used when extracting cookies for this
-            profile (0 for the default account).
-        email: Optional account email to record alongside the index.
-    """
-    context_path = _account_context_path(storage_path)
-    payload: dict[str, Any] = {"authuser": authuser}
-    if email:
-        payload["email"] = email
-
-    def _set_account(data: dict[str, Any]) -> dict[str, Any]:
-        data[_ACCOUNT_CONTEXT_KEY] = payload
-        return data
-
-    # ``recover_from_corrupt=True`` keeps the empty-dict fallback **inside**
-    # the file lock. An outside-the-lock unlink-and-retry would race a
-    # concurrent process that wrote a valid payload between our raise and
-    # our retry, causing us to delete their good write (see PR #465 review).
-    # Account metadata is unrecoverable from corrupt JSON, so silent reset
-    # under the lock is the right behaviour.
-    atomic_update_json(context_path, _set_account, recover_from_corrupt=True)
-
-
-def clear_account_metadata(storage_path: Path | None) -> None:
-    """Remove account metadata from sibling ``context.json`` if present.
-
-    Holds the same sibling ``.lock`` file used by :func:`atomic_update_json`
-    so concurrent ``write_account_metadata`` / context-mutation calls don't
-    lose updates against our clear-and-maybe-delete.
-    """
-    if storage_path is None:
-        return
-    context_path = _account_context_path(storage_path)
-    if not context_path.exists():
-        return
-    lock_path = context_path.with_suffix(context_path.suffix + ".lock")
-    context_path.parent.mkdir(parents=True, exist_ok=True)
-    with FileLock(str(lock_path), timeout=10.0):
-        # Re-check existence under the lock — another writer may have
-        # removed it between the early-return check and the lock acquire.
-        if not context_path.exists():
-            return
-        try:
-            data = json.loads(context_path.read_text(encoding="utf-8"))
-        except (OSError, json.JSONDecodeError) as e:
-            logger.debug("account metadata clear skipped for %s: %s", context_path, e)
-            return
-        if not isinstance(data, dict) or _ACCOUNT_CONTEXT_KEY not in data:
-            return
-        del data[_ACCOUNT_CONTEXT_KEY]
-        if data:
-            # atomic_update_json would re-acquire the lock; use the atomic
-            # write directly since we already hold the lock here.
-            atomic_write_json(context_path, data)
-        else:
-            context_path.unlink()
+    """Enumerate Google accounts visible to the given cookie jar."""
+    return await _auth_account.enumerate_accounts(
+        cookie_jar,
+        max_authuser=max_authuser,
+        poke_session=_poke_session,
+    )
 
 
 def load_auth_from_storage(path: Path | None = None) -> dict[str, str]:
@@ -867,628 +584,6 @@ def load_auth_from_storage(path: Path | None = None) -> dict[str, str]:
     """
     storage_state = _load_storage_state(path)
     return extract_cookies_from_storage(storage_state)
-
-
-_LOCK_CONTENTION_ERRNOS = {errno.EWOULDBLOCK, errno.EAGAIN, errno.EACCES}
-
-
-@contextlib.contextmanager
-def _file_lock(lock_path: Path, *, blocking: bool, log_prefix: str) -> Iterator[str]:
-    """Cross-process exclusive lock on ``lock_path``.
-
-    Yields one of:
-      - ``"held"``  — the lock is held; release it on exit.
-      - ``"contended"`` — non-blocking acquire saw the lock held elsewhere.
-        Only ever yielded when ``blocking=False``.
-      - ``"unavailable"`` — lock infrastructure failed (cannot mkdir, cannot
-        open the sentinel, NFS without flock support). Caller should
-        **fail open** (proceed without coordination) rather than retry forever.
-
-    Wrappers translate this tristate into bool. Distinguishing contention from
-    infrastructure failure matters: a non-blocking caller should **skip** on
-    contention (someone else is rotating) but **proceed** on infrastructure
-    failure (otherwise a read-only auth dir would permanently suppress
-    rotation).
-    """
-    try:
-        lock_path.parent.mkdir(parents=True, exist_ok=True)
-        fd = os.open(lock_path, os.O_RDWR | os.O_CREAT, 0o600)
-    except OSError as exc:
-        # Read-only directory, permission denied, ENOSPC, etc. Yield
-        # "unavailable" so the wrapper can fail open.
-        logger.debug("%s: lock file unavailable %s (%s)", log_prefix, lock_path, exc)
-        yield "unavailable"
-        return
-    locked = False
-    state = "unavailable"
-    try:
-        try:
-            if sys.platform == "win32":
-                import msvcrt
-
-                msvcrt.locking(fd, msvcrt.LK_LOCK if blocking else msvcrt.LK_NBLCK, 1)
-            else:
-                import fcntl
-
-                op = fcntl.LOCK_EX if blocking else fcntl.LOCK_EX | fcntl.LOCK_NB
-                fcntl.flock(fd, op)
-            locked = True
-            state = "held"
-        except OSError as exc:
-            if not blocking and exc.errno in _LOCK_CONTENTION_ERRNOS:
-                # Non-blocking acquire bounced because another process holds
-                # the lock — this is the "skip" signal.
-                state = "contended"
-                logger.debug("%s: lock contended (%s)", log_prefix, exc)
-            else:
-                # NFS without flock, kernel quirk, etc. Caller should fail open.
-                state = "unavailable"
-                logger.debug("%s: lock op unavailable (%s)", log_prefix, exc)
-        yield state
-    finally:
-        if locked:
-            try:
-                if sys.platform == "win32":
-                    import msvcrt
-
-                    msvcrt.locking(fd, msvcrt.LK_UNLCK, 1)
-                else:
-                    import fcntl
-
-                    fcntl.flock(fd, fcntl.LOCK_UN)
-            except OSError as exc:
-                logger.debug("%s: failed to release file lock (%s)", log_prefix, exc)
-        os.close(fd)
-
-
-# Dedupe contract (T7.G7): best-effort under threads, exactly-once on a single
-# event loop. ``_file_lock_exclusive`` below reads ``_FLOCK_UNAVAILABLE_WARNED``
-# and sets it to ``True`` in one synchronous block with no intervening
-# ``await``, so concurrent coroutines on one loop cannot interleave between
-# the check and the set — the warning fires exactly once per process. Under
-# genuine OS threads (out of scope per the documented concurrency contract),
-# duplicate warnings are possible. We accept that rather than serialize a
-# logging side-effect behind a lock for an unsupported configuration.
-#
-# Note: ``functools.lru_cache`` and ``logging.LoggerAdapter`` do NOT solve
-# this — ``lru_cache`` memoizes return values, not the ``logger.warning``
-# side-effect; ``LoggerAdapter`` only rewrites records, it does not filter
-# duplicates.
-_FLOCK_UNAVAILABLE_WARNED = False
-
-
-@contextlib.contextmanager
-def _file_lock_exclusive(lock_path: Path) -> Iterator[None]:
-    """Blocking cross-process exclusive lock on ``lock_path``.
-
-    Multiple Python processes that all save to the same ``storage_state.json``
-    (e.g. a long-running ``NotebookLMClient(keepalive=...)`` worker plus a
-    cron-driven ``notebooklm auth refresh``) would otherwise race on the read-
-    merge-write cycle and lose updates. The lock is held on a sentinel file
-    sibling to the storage file (``.storage_state.json.lock``), since locking
-    the storage file itself would interfere with the atomic temp-rename below.
-
-    The lock is per-process: threads within one process aren't serialized —
-    that's the intra-process ``threading.Lock`` in ``ClientCore``. If the
-    lock can't be acquired (e.g. NFS where flock semantics vary, read-only
-    parent dir, fd exhaustion), the save proceeds anyway; correctness in
-    that mode is best-effort and relies on the snapshot/delta CAS guards in
-    :func:`_merge_cookies_with_snapshot` alone. The first time this
-    fallback fires per process emits a WARNING so operators learn their
-    deployment is running without cross-process coordination.
-    """
-    global _FLOCK_UNAVAILABLE_WARNED
-    with _file_lock(lock_path, blocking=True, log_prefix="save_cookies_to_storage") as state:
-        if state == "unavailable" and not _FLOCK_UNAVAILABLE_WARNED:
-            _FLOCK_UNAVAILABLE_WARNED = True
-            logger.warning(
-                "Cross-process file lock unavailable at %s; cookie saves will "
-                "proceed without cross-process coordination and rely solely on "
-                "snapshot/delta CAS guards. Common causes: NFS without flock "
-                "support, read-only parent directory, fd exhaustion. (Logged "
-                "once per process.)",
-                lock_path,
-            )
-        yield
-
-
-def snapshot_cookie_jar(cookie_jar: httpx.Cookies) -> CookieSnapshot:
-    """Capture an open-time snapshot of an httpx cookie jar.
-
-    Snapshots are the input to the dirty-flag/delta merge in
-    :func:`save_cookies_to_storage`: at save time, only cookies whose
-    in-memory value differs from the snapshot — plus cookies absent from
-    the jar but present in the snapshot (deletions) — are propagated to
-    disk. Cookies the in-process code never touched are left to whatever
-    a sibling process may have written (closes the §3.4.1
-    stale-overwrite-fresh hazard).
-
-    The key shape is path-aware ``(name, domain, path)`` (also closes
-    §3.4.2). Cookies with no name or no domain are skipped — the storage
-    format requires both.
-
-    Args:
-        cookie_jar: The httpx.Cookies object to snapshot.
-
-    Returns:
-        Mapping of ``CookieSnapshotKey -> CookieSnapshotValue`` capturing
-        each cookie's value and the attributes the storage_state schema
-        persists (``expires``, ``secure``, ``httpOnly``).
-    """
-    return {
-        CookieSnapshotKey(cookie.name, cookie.domain, cookie.path or "/"): CookieSnapshotValue(
-            value=cookie.value,
-            expires=cookie.expires,
-            secure=bool(cookie.secure),
-            http_only=_cookie_is_http_only(cookie),
-        )
-        for cookie in cookie_jar.jar
-        if cookie.name and cookie.domain and cookie.value is not None
-    }
-
-
-def _cookie_snapshot_key_variants(key: CookieSnapshotKey) -> set[CookieSnapshotKey]:
-    """Return equivalent host/domain snapshot keys for leading-dot domains.
-
-    Mirrors :func:`_cookie_key_variants` but preserves the path component so
-    storage entries on the same path match snapshot entries regardless of
-    whether ``http.cookiejar`` normalized the domain to a leading dot.
-    """
-    variants = {key}
-    if key.domain.startswith("."):
-        variants.add(CookieSnapshotKey(key.name, key.domain[1:], key.path))
-    else:
-        variants.add(CookieSnapshotKey(key.name, f".{key.domain}", key.path))
-    return variants
-
-
-def _stored_cookie_snapshot_key(stored_cookie: dict[str, Any]) -> CookieSnapshotKey | None:
-    """Build a path-aware snapshot key from a Playwright storage_state cookie."""
-    name = stored_cookie.get("name")
-    domain = stored_cookie.get("domain", "")
-    if not name or not domain:
-        return None
-    path = stored_cookie.get("path") or "/"
-    return CookieSnapshotKey(name, domain, path)
-
-
-def advance_cookie_snapshot_after_save(
-    original_snapshot: CookieSnapshot | None,
-    post_save_snapshot: CookieSnapshot,
-    cas_rejected_keys: frozenset[CookieSnapshotKey],
-) -> CookieSnapshot | None:
-    """Advance save baseline for successful keys while preserving rejected ones.
-
-    A save can partially succeed: one cookie delta may write through while a
-    sibling-process CAS conflict rejects another. Advancing the whole baseline
-    would lose the rejected delta; keeping the whole old baseline would replay
-    already-written deltas and wedge future saves. This helper advances every
-    key to ``post_save_snapshot`` except the CAS-rejected keys, which retain
-    their old baseline value or absence. Rejected keys are matched through
-    leading-dot variants because the merge path can reject a normalized variant
-    of the key captured in ``original_snapshot``.
-    """
-    if original_snapshot is None:
-        return None
-
-    advanced = dict(post_save_snapshot)
-    for key in cas_rejected_keys:
-        original_key = next(
-            (
-                variant
-                for variant in _cookie_snapshot_key_variants(key)
-                if variant in original_snapshot
-            ),
-            None,
-        )
-        for variant in _cookie_snapshot_key_variants(key):
-            advanced.pop(variant, None)
-        if original_key is not None:
-            advanced[original_key] = original_snapshot[original_key]
-    return advanced
-
-
-def _cookie_save_return(
-    result: CookieSaveResult, *, return_result: bool
-) -> bool | CookieSaveResult:
-    """Return either the detailed save result or its public bool projection."""
-    return result if return_result else result.ok
-
-
-def save_cookies_to_storage(
-    cookie_jar: httpx.Cookies,
-    path: Path | None = None,
-    *,
-    original_snapshot: CookieSnapshot | None = None,
-    return_result: bool = False,
-) -> bool | CookieSaveResult:
-    """Save an updated httpx.Cookies jar back to Playwright storage_state.json.
-
-    This ensures that when Google issues short-lived token refreshes (e.g.
-    during 302 redirects to accounts.google.com), those updated cookies are
-    serialized back to disk so the session remains valid across CLI invocations.
-
-    If auth was loaded from an environment variable (no file), this is a no-op.
-
-    Cross-process safety: the read-merge-write cycle is wrapped in an OS-level
-    file lock (``.storage_state.json.lock``) so concurrent writers from
-    different Python processes (e.g. an in-process ``NotebookLMClient`` keepalive
-    plus a cron-driven ``notebooklm auth refresh``) serialize cleanly rather
-    than tearing or losing updates.
-
-    Two merge modes:
-
-    - **Legacy (``original_snapshot=None``)**: every in-memory cookie whose
-      value differs from disk wins. Vulnerable to the stale-overwrite-fresh
-      race documented in ``docs/auth-keepalive.md`` §3.4.1 and emits a
-      ``DeprecationWarning``. Kept only as a public-API back-compat shim
-      for callers outside this repo; every first-party caller passes
-      ``original_snapshot``.
-    - **Snapshot/delta (``original_snapshot`` provided)**: only cookies
-      whose in-memory persisted tuple differs from the snapshot are written, and
-      cookies present in the snapshot but no longer in the jar are
-      deleted from disk. Cookies the in-process code never touched are
-      left untouched on disk so a sibling-process write survives.
-      Path-aware ``(name, domain, path)`` keys are used here (also closes
-      §3.4.2).
-
-    Args:
-        cookie_jar: The httpx.Cookies object containing the latest cookies.
-        path: Path to storage_state.json. If None, cookie sync is skipped.
-        original_snapshot: Open-time snapshot from
-            :func:`snapshot_cookie_jar`. When provided, only deltas and
-            deletions relative to the snapshot are persisted.
-        return_result: Internal escape hatch for callers that need CAS-rejected
-            keys to maintain a per-cookie baseline. Public callers should use
-            the default bool return.
-
-    Returns:
-        ``True`` if the disk state now reflects the caller's intent (write
-        succeeded, was a successful no-op, or the call was a deliberate skip
-        because auth was loaded from an env var). ``False`` if an I/O error
-        prevented the save or a CAS guard preserved a sibling-process write.
-        With ``return_result=True``, callers can inspect CAS-rejected keys and
-        advance their baseline for the keys that did write through.
-    """
-    if (
-        not path
-        and "NOTEBOOKLM_AUTH_JSON" in os.environ
-        and os.environ["NOTEBOOKLM_AUTH_JSON"].strip()
-    ):
-        logger.debug("Skipping cookie sync: Auth loaded from NOTEBOOKLM_AUTH_JSON env var")
-        return _cookie_save_return(CookieSaveResult(True), return_result=return_result)
-
-    if not path:
-        logger.debug("Skipping cookie sync: No storage file path available")
-        return _cookie_save_return(CookieSaveResult(True), return_result=return_result)
-
-    if original_snapshot is None:
-        warnings.warn(
-            "save_cookies_to_storage called without original_snapshot; the "
-            "legacy full-merge path is vulnerable to the stale-overwrite-fresh "
-            "race (docs/auth-keepalive.md §3.4.1). Pass an original_snapshot "
-            "captured via snapshot_cookie_jar() at jar-open time.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-
-    lock_path = path.with_name(f".{path.name}.lock")
-    with _file_lock_exclusive(lock_path):
-        if not path.exists():
-            logger.debug("Skipping cookie sync: Storage file not found at %s", path)
-            return _cookie_save_return(CookieSaveResult(False), return_result=return_result)
-
-        try:
-            storage_data = json.loads(path.read_text(encoding="utf-8"))
-        except Exception as e:
-            logger.warning("Failed to read storage state for cookie sync: %s", e)
-            return _cookie_save_return(CookieSaveResult(False), return_result=return_result)
-
-        if not isinstance(storage_data, dict) or "cookies" not in storage_data:
-            logger.warning(
-                "storage_state at %s lacks 'cookies' key; rotated cookies will not be persisted",
-                path,
-            )
-            return _cookie_save_return(CookieSaveResult(False), return_result=return_result)
-
-        if original_snapshot is None:
-            updated_count = _merge_cookies_legacy(cookie_jar, storage_data)
-            cas_rejected_keys: frozenset[CookieSnapshotKey] = frozenset()
-        else:
-            updated_count, cas_rejected_keys = _merge_cookies_with_snapshot(
-                cookie_jar, storage_data, original_snapshot
-            )
-
-        if updated_count == 0:
-            # A CAS rejection with no other successful work means disk does
-            # not reflect our intent; the caller must not advance baseline.
-            return _cookie_save_return(
-                CookieSaveResult(not cas_rejected_keys, cas_rejected_keys),
-                return_result=return_result,
-            )
-
-        try:
-            atomic_write_json(path, storage_data)
-            logger.debug("Successfully synced %d refreshed cookies to %s", updated_count, path)
-            # Even on a successful disk write, if any CAS arm rejected work,
-            # disk diverges from ``post`` for at least one key — caller must
-            # not advance baseline.
-            return _cookie_save_return(
-                CookieSaveResult(not cas_rejected_keys, cas_rejected_keys),
-                return_result=return_result,
-            )
-        except Exception as e:
-            logger.warning("Failed to write updated cookies to %s: %s", path, e)
-            return _cookie_save_return(CookieSaveResult(False), return_result=return_result)
-
-
-def _merge_cookies_legacy(cookie_jar: httpx.Cookies, storage_data: dict[str, Any]) -> int:
-    """Legacy merge: trust in-memory whenever it differs from disk.
-
-    Vulnerable to the stale-overwrite-fresh race (§3.4.1). Kept only for
-    callers that have not yet opted into snapshot semantics. New callers
-    must pass ``original_snapshot`` to :func:`save_cookies_to_storage`.
-
-    Returns:
-        Number of cookie entries added or modified in ``storage_data``.
-    """
-    cookies_by_key: dict[CookieKey, Any] = {
-        (cookie.name, cookie.domain, cookie.path or "/"): cookie
-        for cookie in cookie_jar.jar
-        if cookie.name and cookie.domain and _is_allowed_cookie_domain(cookie.domain)
-    }
-
-    updated_count = 0
-    stored_keys: set[CookieKey] = set()
-    for stored_cookie in storage_data["cookies"]:
-        name = stored_cookie.get("name")
-        domain = stored_cookie.get("domain", "")
-        if not name or not domain:
-            continue
-
-        key: CookieKey = (name, domain, stored_cookie.get("path") or "/")
-        stored_keys.update(_cookie_key_variants(key))
-        refreshed_cookie = _find_cookie_for_storage(cookies_by_key, key, stored_cookie.get("value"))
-        if refreshed_cookie is None:
-            continue
-
-        new_expires = refreshed_cookie.expires if refreshed_cookie.expires is not None else -1
-        changed = (
-            stored_cookie.get("value") != refreshed_cookie.value
-            or stored_cookie.get("expires") != new_expires
-        )
-        if changed:
-            stored_cookie["value"] = refreshed_cookie.value
-            stored_cookie["expires"] = new_expires
-            # Normalize present-but-empty ``"path": ""`` to ``"/"`` so the row
-            # we write matches the path normalization used to build the
-            # identity key one block up (and used by every loader). Without
-            # the trailing ``or "/"`` an on-disk row with ``"path": ""`` would
-            # survive across save cycles while every other code path treats
-            # it as ``"/"``.
-            stored_cookie["path"] = refreshed_cookie.path or stored_cookie.get("path") or "/"
-            stored_cookie["secure"] = refreshed_cookie.secure
-            stored_cookie["httpOnly"] = _cookie_is_http_only(refreshed_cookie)
-            updated_count += 1
-
-    for key, cookie in cookies_by_key.items():
-        if key in stored_keys:
-            continue
-        storage_data["cookies"].append(_cookie_to_storage_state(cookie))
-        updated_count += 1
-
-    return updated_count
-
-
-def _merge_cookies_with_snapshot(
-    cookie_jar: httpx.Cookies,
-    storage_data: dict[str, Any],
-    original_snapshot: CookieSnapshot,
-) -> tuple[int, frozenset[CookieSnapshotKey]]:
-    """Snapshot/delta merge: write only what this process actually changed.
-
-    Closes §3.4.1 (stale-overwrite-fresh) and §3.4.2 (path collapse):
-
-    - **Deltas (CAS-guarded for keys in the snapshot)**: cookies in the
-      jar whose snapshot tuple (``value, expires, secure, http_only``)
-      differs from ``original_snapshot`` are written to disk **only if**
-      the on-disk value still matches the snapshot value. If disk has
-      rotated since open time, a sibling process has written it; we
-      preserve their write rather than clobber it with our local
-      rotation. New cookies acquired during the session are written only
-      when no same-key storage row exists yet; an existing row means a
-      sibling acquired the same cookie first. Comparing the full snapshot
-      tuple keeps attribute-only refreshes (same value, new ``expires``)
-      flowing to disk, but CAS remains value-only because attribute-only
-      sibling drift is routine session metadata and should not wedge later
-      value rotations.
-    - **Deletions (CAS-guarded)**: a key present in the snapshot but
-      absent from the jar is dropped from disk **only if** the on-disk
-      value still matches the snapshot value — symmetric with the
-      value-update CAS above. An ``Max-Age=0`` that evicted our
-      locally-expired copy must not erase the sibling's freshly-issued
-      replacement.
-    - **Untouched**: cookies in the jar whose tuple matches the snapshot
-      are not written, so a sibling-process write to the same key
-      survives. Cookies on disk that are not in the snapshot are also
-      left alone (they belong to a sibling process or another path).
-
-    Args:
-        cookie_jar: Current in-memory cookie jar.
-        storage_data: Mutable storage_state.json dict (modified in place).
-        original_snapshot: Open-time snapshot of the same jar.
-
-    Returns:
-        Tuple of ``(updated_count, cas_rejected_keys)``:
-
-        - ``updated_count``: cookie entries added, modified, or removed
-          (drives whether the temp-write step runs).
-        - ``cas_rejected_keys``: keys whose CAS check rejected a delta or
-          deletion. Caller uses this to advance the baseline only for keys
-          that were actually written or already matched.
-    """
-    current_snapshot = snapshot_cookie_jar(cookie_jar)
-
-    # Path-aware index of jar cookies for delta application. Restricting to
-    # _is_allowed_cookie_domain matches the legacy save's allowlist gate so
-    # this PR doesn't inadvertently widen the persisted-domain set.
-    # Filter ``cookie.value is not None`` to mirror ``snapshot_cookie_jar``: a
-    # value-less cookie is treated as a deletion (absent from this index, absent
-    # from ``current_snapshot``) rather than a delta that would write ``null``
-    # to disk.
-    cookies_by_snapshot_key = {
-        CookieSnapshotKey(cookie.name, cookie.domain, cookie.path or "/"): cookie
-        for cookie in cookie_jar.jar
-        if (
-            cookie.name
-            and cookie.domain
-            and cookie.value is not None
-            and _is_allowed_cookie_domain(cookie.domain)
-        )
-    }
-
-    deltas: dict[CookieSnapshotKey, Any] = {}
-    for snapshot_key, cookie in cookies_by_snapshot_key.items():
-        if original_snapshot.get(snapshot_key) != current_snapshot.get(snapshot_key):
-            deltas[snapshot_key] = cookie
-
-    deletion_candidates: set[CookieSnapshotKey] = {
-        snapshot_key
-        for snapshot_key in original_snapshot
-        if snapshot_key not in current_snapshot
-        # Only delete cookies the merge would otherwise be allowed to write.
-        # Snapshot may include sibling-product domains the allowlist filters
-        # out at write time; treating those as deletions would silently drop
-        # disk entries we never persisted to begin with.
-        and _is_allowed_cookie_domain(snapshot_key.domain)
-    }
-
-    updated_count = 0
-    cas_rejected_keys: set[CookieSnapshotKey] = set()
-
-    # Apply deltas + deletions to the existing storage entries in place.
-    new_cookies: list[dict[str, Any]] = []
-    matched_delta_keys: set[CookieSnapshotKey] = set()
-    for stored_cookie in storage_data["cookies"]:
-        stored_key = _stored_cookie_snapshot_key(stored_cookie)
-        if stored_key is None:
-            new_cookies.append(stored_cookie)
-            continue
-
-        # Find the delta (or deletion) that maps to this stored entry.
-        # Match leading-dot domain variants so e.g. snapshot
-        # ``.accounts.google.com`` lines up with stored ``accounts.google.com``.
-        # A delta wins over a deletion: if the same stored entry matches
-        # both (which can happen when httpx normalized one variant), we
-        # prefer to update rather than drop, because dropping would lose
-        # the rotation we just applied.
-        matched_delta_cookie = None
-        matched_delta_key: CookieSnapshotKey | None = None
-        for variant in _cookie_snapshot_key_variants(stored_key):
-            if variant in deltas:
-                matched_delta_cookie = deltas[variant]
-                matched_delta_key = variant
-                break
-
-        if matched_delta_cookie is not None:
-            if matched_delta_key is None:  # pragma: no cover - loop invariant
-                raise RuntimeError("matched_delta_cookie set without matched_delta_key")
-            # CAS-guard for value updates: if our snapshot had this key in any
-            # leading-dot variant and disk's current value differs from the
-            # snapshot value, a sibling process has rewritten the row between
-            # our open and our save. Preserve their write rather than clobber,
-            # unless disk has already converged to our current value; in that
-            # case the save intent is satisfied and the caller may advance its
-            # baseline.
-            # Variant-aware lookup mirrors the delta match above: if the snapshot
-            # was keyed on ``accounts.google.com`` but the matched delta key is
-            # the leading-dot variant, a plain ``.get(matched_delta_key)`` would
-            # miss the entry and silently bypass the CAS.
-            snapshot_entry = next(
-                (
-                    original_snapshot[variant]
-                    for variant in _cookie_snapshot_key_variants(matched_delta_key)
-                    if variant in original_snapshot
-                ),
-                None,
-            )
-            stored_value = stored_cookie.get("value")
-            if (
-                snapshot_entry is not None
-                and stored_value != snapshot_entry.value
-                and stored_value != matched_delta_cookie.value
-            ):
-                logger.debug(
-                    "Skipped CAS-guarded value update of %s on %s: disk value "
-                    "differs from snapshot (sibling write preserved)",
-                    matched_delta_key.name,
-                    matched_delta_key.domain,
-                )
-                cas_rejected_keys.add(matched_delta_key)
-                matched_delta_keys.add(matched_delta_key)
-                new_cookies.append(stored_cookie)
-                continue
-            if snapshot_entry is None and stored_value != matched_delta_cookie.value:
-                logger.debug(
-                    "Skipped CAS-guarded value update of new cookie %s on %s: "
-                    "disk row already exists (sibling write preserved)",
-                    matched_delta_key.name,
-                    matched_delta_key.domain,
-                )
-                cas_rejected_keys.add(matched_delta_key)
-                matched_delta_keys.add(matched_delta_key)
-                new_cookies.append(stored_cookie)
-                continue
-            new_expires = (
-                matched_delta_cookie.expires if matched_delta_cookie.expires is not None else -1
-            )
-            stored_cookie["value"] = matched_delta_cookie.value
-            stored_cookie["expires"] = new_expires
-            # Mirror :func:`_merge_cookies_legacy`: ``or "/"`` normalizes a
-            # present-but-empty ``"path": ""`` so the written row matches the
-            # path normalization used by the identity key and every loader.
-            stored_cookie["path"] = matched_delta_cookie.path or stored_cookie.get("path") or "/"
-            stored_cookie["secure"] = matched_delta_cookie.secure
-            stored_cookie["httpOnly"] = _cookie_is_http_only(matched_delta_cookie)
-            matched_delta_keys.add(matched_delta_key)
-            updated_count += 1
-            new_cookies.append(stored_cookie)
-            continue
-
-        deletion_match = next(
-            (
-                variant
-                for variant in _cookie_snapshot_key_variants(stored_key)
-                if variant in deletion_candidates
-            ),
-            None,
-        )
-        if deletion_match is not None:
-            # CAS-guard: only drop the disk row if its value still matches
-            # what we observed at snapshot time. A sibling process may have
-            # rewritten this key between our open and our save; clobbering
-            # their fresh value with our local eviction would resurrect the
-            # exact stale-overwrite-fresh hazard the snapshot path exists
-            # to close (just inverted — deletion-of-fresh instead of
-            # value-write-of-stale).
-            snapshot_value = original_snapshot[deletion_match].value
-            if stored_cookie.get("value") == snapshot_value:
-                updated_count += 1
-                continue  # drop the entry from disk
-            cas_rejected_keys.add(deletion_match)
-
-        new_cookies.append(stored_cookie)
-
-    # Append delta cookies that didn't match any existing storage entry
-    # (genuinely new cookies acquired during the session).
-    for snapshot_key, cookie in deltas.items():
-        if snapshot_key in matched_delta_keys:
-            continue
-        new_cookies.append(_cookie_to_storage_state(cookie))
-        updated_count += 1
-
-    storage_data["cookies"] = new_cookies
-    return updated_count, frozenset(cas_rejected_keys)
 
 
 NOTEBOOKLM_REFRESH_CMD_ENV = "NOTEBOOKLM_REFRESH_CMD"
